fix: IndexError in Line
For TCP Retransmission/Fast Retransmission frames,
it is strange that TLS dissector is called normally,
but does not show in the dissector tree. Moreover,
for HTTP/2 reassemble info, these TLS frames are
shown correctly.

It causes some TLS segments in the upper layer does
not show in the lower layer, which might cause
IndexError.

Current fix is awkward, which skips the wrong Line,
and we remove those problematic capture files...