fix: wrong IQR computation
np.percentile uses 'linear' method to select the
inter-percentile value, which does not correspond
to IQR requirements. Changing to 'midpoint' solves
this issue. However, different version of Numpy uses
different params for it, e.g., 'interpolation' and
'method'.

To avoid version dependency, we manually computes
the IQR, and make it a util function for code reuse.