feat: overwrite layer name for Trojan
In Wireshark C plugins, it is hard to modify the
protocol register names, which leads that in PyShark,
which reads PDML, to consider Trojan layers as TLS
ones.

One quick fix is through showname, which allows
modifying in C plugin backend. However, early implementations
such as seq_filter, CellExtractor and so on, relies
heavily on layer_name of XmlLayer objects.

To restrict the side affects of changing the source
code, we confine the Trojan compatibility in
layer_extractor. It modifies the layer_name of both
Trojan and Fake Trojan (for display filter) to the
proper names, such that the caller of layer_extractor
does not need to worry about whether to use layer_name
or layer_showname.