fix: mistake TLS for Trojan packets
Upon the introduction of layer_rename, if we don't check if the target
is actually a Trojan packet, it may incorrect recognize TLS records to
Trojan records.

For packet-level fix, i.e., layer_extractor, it is enough to check if
the upper/lower protocol is Trojan to notify the function whether this
packet might be a Trojan packet.

However, for stream-level fix, i.e., the get_adjacent_protocol_reassemble_info
and get_reassemble_info since the tunnel TLS handshake is completely a
normal TLS handshake, it is the caller who is responsible to notify the
callee that the stream to analyze is a Trojan stream.

We currently only add the tunnel_tls variable as the notification for
stream-level fix. In the future, we may also add it to packet-level fix
for better extendability.

Such fix should also be applied to all TLS tunnel based proxy protocols
like VLess.